+++ src/create-svg.ts
@@ -5,9 +5,6 @@ import * as pie from './create-pie-language';
 import * as radar from './create-radar-contrib';
 import * as type from './type';
 
-const fgcolor = '#00000f';
-const bgcolor = '#ffffff';
-const strongColor = '#111133';
 const width = 1280;
 const height = 850;
 
@@ -15,7 +12,7 @@ const toIsoDate = (date: Date) => date.toISOString().substring(0, 10);
 
 export const createSvg = (
     userInfo: type.UserInfo,
-    seasonMode: type.SeasonMode,
+    settings: type.Settings,
     isAnimate: boolean
 ): string => {
     const fakeDom = new JSDOM(
@@ -39,7 +36,7 @@ export const createSvg = (
         .attr('y', 0)
         .attr('width', width)
         .attr('height', height)
-        .attr('fill', bgcolor);
+        .attr('fill', settings.backgroundColor);
 
     contrib.create3DContrib(
         svg,
@@ -48,7 +45,7 @@ export const createSvg = (
         0,
         width,
         height,
-        seasonMode,
+        settings,
         isAnimate
     );
 
@@ -63,6 +60,7 @@ export const createSvg = (
         70,
         radarWidth,
         radarHeight,
+        settings,
         isAnimate
     );
 
@@ -76,6 +74,7 @@ export const createSvg = (
         height - pieHeight - 70,
         pieWidth,
         pieHeight,
+        settings,
         isAnimate
     );
 
@@ -91,9 +90,14 @@ export const createSvg = (
         .attr('x', positionXContrib)
         .attr('y', positionYContrib)
         .attr('text-anchor', 'end')
-        .text(userInfo.totalContributions.toLocaleString())
-        // .text(userInfo.totalContributions.toLocaleString().replace(',', ' ')) // for SI
-        .attr('fill', strongColor);
+        .text(
+            userInfo.totalContributions < 1000
+                ? userInfo.totalContributions
+                : '999+'
+        )
+        .attr('fill', settings.strongColor)
+        .append('title')
+        .text(userInfo.totalContributions);
 
     group
         .append('text')
@@ -103,7 +107,7 @@ export const createSvg = (
         .attr('text-anchor', 'start')
         .attr('text-anchor', 'start')
         .text('contributions')
-        .attr('fill', fgcolor);
+        .attr('fill', settings.foregroundColor);
 
     const positionXStar = (width * 5) / 10;
     const positionYStar = positionYContrib;
@@ -121,7 +125,7 @@ export const createSvg = (
             'd',
             'M8 .25a.75.75 0 01.673.418l1.882 3.815 4.21.612a.75.75 0 01.416 1.279l-3.046 2.97.719 4.192a.75.75 0 01-1.088.791L8 12.347l-3.766 1.98a.75.75 0 01-1.088-.79l.72-4.194L.818 6.374a.75.75 0 01.416-1.28l4.21-.611L7.327.668A.75.75 0 018 .25zm0 2.445L6.615 5.5a.75.75 0 01-.564.41l-3.097.45 2.24 2.184a.75.75 0 01.216.664l-.528 3.084 2.769-1.456a.75.75 0 01.698 0l2.77 1.456-.53-3.084a.75.75 0 01.216-.664l2.24-2.183-3.096-.45a.75.75 0 01-.564-.41L8 2.694v.001z'
         )
-        .attr('fill', fgcolor);
+        .attr('fill', settings.foregroundColor);
 
     group
         .append('text')
@@ -135,7 +139,9 @@ export const createSvg = (
                 ? userInfo.totalStargazerCount
                 : '999+'
         )
-        .attr('fill', fgcolor);
+        .attr('fill', settings.foregroundColor)
+        .append('title')
+        .text(userInfo.totalStargazerCount);
 
     const positionXFork = (width * 6) / 10;
     const positionYFork = positionYContrib;
@@ -153,7 +159,7 @@ export const createSvg = (
             'd',
             'M5 3.25a.75.75 0 11-1.5 0 .75.75 0 011.5 0zm0 2.122a2.25 2.25 0 10-1.5 0v.878A2.25 2.25 0 005.75 8.5h1.5v2.128a2.251 2.251 0 101.5 0V8.5h1.5a2.25 2.25 0 002.25-2.25v-.878a2.25 2.25 0 10-1.5 0v.878a.75.75 0 01-.75.75h-4.5A.75.75 0 015 6.25v-.878zm3.75 7.378a.75.75 0 11-1.5 0 .75.75 0 011.5 0zm3-8.75a.75.75 0 100-1.5.75.75 0 000 1.5z'
         )
-        .attr('fill', fgcolor);
+        .attr('fill', settings.foregroundColor);
 
     group
         .append('text')
@@ -163,7 +169,9 @@ export const createSvg = (
         .attr('y', positionYFork)
         .attr('text-anchor', 'start')
         .text(userInfo.totalForkCount < 1000 ? userInfo.totalForkCount : '999+')
-        .attr('fill', fgcolor);
+        .attr('fill', settings.foregroundColor)
+        .append('title')
+        .text(userInfo.totalForkCount);
 
     // ISO 8601 format
     const startDate = userInfo.contributionCalendar[0].date;
@@ -180,7 +188,7 @@ export const createSvg = (
         .attr('dominant-baseline', 'hanging')
         .attr('text-anchor', 'end')
         .text(period)
-        .attr('fill', 'gray');
+        .attr('fill', settings.weakColor);
 
     return container.html();
 };