+++ src/type.ts
@@ -31,4 +31,46 @@ export type ContributionLevel =
     | 'THIRD_QUARTILE'
     | 'FOURTH_QUARTILE';
 
-export type SeasonMode = 'season' | 'green' | 'halloween';
+export interface NormalColorSettings {
+    type: 'normal';
+    backgroundColor: string;
+    foregroundColor: string;
+    strongColor: string;
+    weakColor: string;
+    radarColor: string;
+
+    contribColors: [string, string, string, string, string];
+}
+
+export interface SeasonColorSettings {
+    type: 'season';
+    backgroundColor: string;
+    foregroundColor: string;
+    strongColor: string;
+    weakColor: string;
+    radarColor: string;
+
+    contribColors1: [string, string, string, string, string];
+    contribColors2: [string, string, string, string, string];
+    contribColors3: [string, string, string, string, string];
+    contribColors4: [string, string, string, string, string];
+}
+
+export interface RainbowColorSettings {
+    type: 'rainbow';
+    backgroundColor: string;
+    foregroundColor: string;
+    strongColor: string;
+    weakColor: string;
+    radarColor: string;
+
+    saturation: number;
+    contribLightness: [string, string, string, string, string];
+    duration: string; // ex. '10s'
+    hueRatio: number; // hue per weeks
+}
+
+export type Settings =
+    | NormalColorSettings
+    | SeasonColorSettings
+    | RainbowColorSettings;